refactor: websocket에 필요한 entity & dto 수정
dto(response,request) 와 entity 정리
