+++ chatting-api/src/main/java/semicolon/viewtist/controller/LiveChatController.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import semicolon.viewtis.chatting.dto.ChatRoomDto;
-import semicolon.viewtis.chatting.form.ChatRoomForm;
+import semicolon.viewtist.chatting.dto.ChatRoomDto;
+import semicolon.viewtist.chatting.form.ChatRoomForm;
 import semicolon.viewtist.service.LiveChatService;
 
 @RestController
