+++ chatting-api/src/main/java/semicolon/viewtist/exception/ChattingException.java
@@ -1,6 +1,7 @@
 package semicolon.viewtist.exception;
 
 import lombok.Getter;
+import semicolon.viewtist.global.exception.ErrorCode;
 
 @Getter
 public class ChattingException extends RuntimeException{
