+++ domain/build.gradle
@@ -7,6 +7,4 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'javax.validation:validation-api:2.0.1.Final'
-    implementation 'org.springframework.data:spring-data-envers'
-
 }
