+++ domain/src/main/java/semicolon/viewtist/chatting/dto/ChatRoomDto.java
@@ -1,11 +1,11 @@
-package semicolon.viewtis.chatting.dto;
+package semicolon.viewtist.chatting.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import semicolon.viewtis.chatting.entity.ChatRoom;
+import semicolon.viewtist.chatting.entity.ChatRoom;
 
 @Builder
 @Getter
