+++ domain/src/main/java/semicolon/viewtist/chatting/entity/ChatRoom.java
@@ -1,4 +1,4 @@
-package semicolon.viewtis.chatting.entity;
+package semicolon.viewtist.chatting.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -8,16 +8,14 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.envers.AuditOverride;
-import semicolon.viewtis.chatting.form.ChatRoomForm;
-import semicolon.viewtis.core.global.BaseTimeEntity;
+import semicolon.viewtist.chatting.form.ChatRoomForm;
+import semicolon.viewtist.global.entitiy.BaseTimeEntity;
 
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
 @Entity
 @Builder
-@AuditOverride(forClass = BaseTimeEntity.class)
 public class ChatRoom extends BaseTimeEntity {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
