+++ domain/src/main/java/semicolon/viewtist/chatting/repository/ChatMessageRepository.java
@@ -1,8 +1,8 @@
-package semicolon.viewtis.chatting.repository;
+package semicolon.viewtist.chatting.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import semicolon.viewtis.chatting.entity.ChatMessage;
+import semicolon.viewtist.chatting.entity.ChatMessage;
 @Repository
 public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> {
