+++ domain/src/main/java/semicolon/viewtist/chatting/repository/ChatRoomRepository.java
@@ -1,10 +1,10 @@
-package semicolon.viewtis.chatting.repository;
+package semicolon.viewtist.chatting.repository;
 
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import semicolon.viewtis.chatting.entity.ChatRoom;
+import semicolon.viewtist.chatting.entity.ChatRoom;
 @Repository
 public interface ChatRoomRepository extends JpaRepository<ChatRoom, Long> {
   Optional <ChatRoom> findByStreamKey(String streamKey);
